Consolidate SkipMixin imports through common-compat layer
Replace version-conditional SkipMixin imports in branch.py and
test_skipmixin.py with the common.compat.sdk layer. Add the standard
provider's skipmixin as a middle fallback path in the compat import
chain for consistency with common/compat/standard/utils.py.

Also route SkipMixin and XComArg in common/sql through the compat
layer instead of importing directly from airflow.models / airflow.

Bump common-compat dependency to >=1.14.0.
