+++ providers/standard/src/airflow/providers/standard/operators/branch.py
@@ -22,12 +22,8 @@ from __future__ import annotations
 from collections.abc import Iterable
 from typing import TYPE_CHECKING
 
-from airflow.providers.standard.version_compat import AIRFLOW_V_3_0_PLUS, BaseOperator
-
-if AIRFLOW_V_3_0_PLUS:
-    from airflow.providers.standard.utils.skipmixin import SkipMixin
-else:
-    from airflow.models.skipmixin import SkipMixin  # type: ignore[no-redef]
+from airflow.providers.common.compat.sdk import SkipMixin
+from airflow.providers.standard.version_compat import BaseOperator
 
 if TYPE_CHECKING:
     from airflow.providers.common.compat.sdk import Context
